Safari-609.1.11.1/LayoutTests/inspector/runtime/evaluate-emulateUserGesture-userIsInteracting-expected.txt

Tests for Runtime emulateUserGesture option.


== Running test suite: Runtime.evaluate.emulateUserGesture.userIsInteracting
-- Running test case: EvaluateWithoutEmulatingUserGesture.userIsInteracting
User is NOT Interacting
-- Running test case: EvaluateWithEmulatingUserGesture.userIsInteracting
User is Interacting
